Implemented 100x data sampling for bme280. Wrote code for writing data to sd card, tested and fully working.

+++ SolarWeatherStation/src/bme280_temp_humi_pres.hpp
@@ -9,9 +9,9 @@
 void bme280_setup(Adafruit_BME280 &bme280);
 
 // Returns readings from the BME280 sensor
-int bme280_get_temperature(Adafruit_BME280 &bme280);
-int bme280_get_humidity(Adafruit_BME280 &bme280);
-int bme280_get_pressure(Adafruit_BME280 &bme280);
+float bme280_get_temperature(Adafruit_BME280 &bme280);
+float bme280_get_humidity(Adafruit_BME280 &bme280);
+float bme280_get_pressure(Adafruit_BME280 &bme280);
 
 // Change sensor mode
 void bme280_sleep_mode();
